server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
email-validator==2.1.0
httpx==0.25.2
Pillow>=10.0.0
aiofiles==23.2.1
